[IMP] web: convert user tests to hoot task-id: 3705027
